Separated player object functions into player.js
Separated the player object functions into its own file, player.js. Additionally changed
the canvas to have a width of 1000px and a height of 600 px.
