+++ index.html
@@ -6,7 +6,7 @@
   <link rel="stylesheet" href="/lib/index.css" media="screen" charset="utf-8">
 </head>
 <body>
-  <canvas tabindex='1' id='game' width='200px' height='200px'></canvas>
+  <canvas tabindex='1' id='game' width='1000px' height='600px'></canvas>
   <script src="main.bundle.js"></script>
 </body>
 </html>
